requirements.txt
Django==4.2
django-admin-sortable2==2.1.8
django-tinymce==3.6.1
Pillow==9.5.0
python-dotenv==1.0.0